Import existing Android project
This took a ton of messing around with Gradle stuff to get this working. Part of it was woring with the Kotlin DSL, which I haven't done much with before. Some of it was also trying to bring over just the relevant parts from the eisting configuration into the new and reconciling issues here and there.

I probably could have done better about breaking this apart, but I'm kind of mentally burnt from messing with htis all day.