Mak platform-specific CompositionPlayers
I'm pretty annoyed by how this turned out. The reason I neeed to do this is that the different versions of JFugue expected different `Sequence` types - since that type was overwritten on Android.

The code is effectively the same for both, just with different imports.

There's probably a clean way to share this code, but I'm more interested in just getting playback working right now.